termohigrometro/backend/requirements.txt
dash
dash-bootstrap-components
pandas
plotly
pyserial